requirements.txt
absl-py==2.3.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
astroid==4.0.1
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
black==25.9.0
bleach==6.3.0
blinker==1.9.0
box2d-py==2.3.5
brax==0.13.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cheroot==11.1.2
chex==0.1.91
click==8.3.0
cloudpickle==3.1.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
datetime==5.5
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
distrax==0.1.7
dm-env==1.6
dm-tree==0.1.9
docstring-parser==0.17.0
dotmap==1.3.30
etils==1.13.0
evosax==0.2.0
executing==2.2.1
farama-notifications==0.0.4
fastjsonschema==2.21.2
filelock==3.20.0
flake8==7.3.0
flashbax==0.1.3
flask==3.1.2
flask-cors==6.0.1
flax==0.12.0
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.9.0
gast==0.6.0
gcsfs==2025.9.0
gitdb==4.0.12
gitpython==3.1.45
glfw==2.10.0
google-api-core==2.28.1
google-auth==2.43.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.5.0
google-cloud-storage==3.5.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.72.0
gviz-api==1.10.0
gymnasium==1.2.1
gymnax==0.0.9
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.14.0
idna==3.11
imageio==2.37.0
importlib-resources==6.5.2
iniconfig==2.3.0
ipykernel==7.1.0
ipympl==0.9.8
ipython==9.6.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
isort==7.0.0
itsdangerous==2.2.0
jaraco-functools==4.3.0
jax==0.8.0
jax-cuda12-pjrt==0.8.0
jax-cuda12-plugin==0.8.0
jaxlib==0.8.0
jaxopt==0.8.5
jedi==0.19.2
jinja2==3.1.6
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-client==8.6.3
jupyter-core==5.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter-server==2.17.0
jupyter-server-terminals==0.5.3
jupyterlab==4.4.10
jupyterlab-pygments==0.3.0
jupyterlab-server==2.28.0
jupyterlab-widgets==3.0.15
kiwisolver==1.4.9
lark==1.3.1
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.4
ml-collections==1.1.0
ml-dtypes==0.5.3
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
mujoco==3.3.7
mujoco-mjx==3.3.7
multidict==6.7.0
mypy-extensions==1.1.0
narwhals==2.10.2
navix==0.7.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.7
notebook-shim==0.2.4
numpy==2.3.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
opt-einsum==3.4.0
optax==0.2.6
orbax-checkpoint==0.11.26
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==12.0.0
pip==25.3
platformdirs==4.5.0
plotly==6.4.0
pluggy==1.6.0
prometheus-client==0.23.1
prompt-toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.2
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.3
pydantic-core==2.41.4
pyflakes==3.4.0
pygame==2.6.1
pygments==2.19.2
pylint==4.0.2
pyopengl==3.1.10
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytokens==0.2.0
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rlax==0.1.7
rpds-py==0.28.0
rsa==4.9.1
scipy==1.16.3
seaborn==0.13.2
send2trash==1.8.3
sentry-sdk==2.42.1
setuptools==80.9.0
setuptools-scm==9.2.2
shtab==1.7.2
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
swig==4.4.0
sympy==1.14.0
tensorboardx==2.6.4
tensorflow-probability==0.25.0
tensorstore==0.1.78
terminado==0.18.1
tfp-nightly==0.26.0.dev20251027
tinycss2==1.4.0
tomlkit==0.13.3
toolz==1.1.0
torch==2.9.0
torchvision==0.24.0
tornado==6.5.2
traitlets==5.14.3
treescope==0.1.10
trimesh==4.9.0
triton==3.5.0
typeguard==4.4.4
typing-extensions==4.15.0
typing-inspection==0.4.2
tyro==0.9.35
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
wandb==0.22.2
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==2.0.0
xprof==2.21.1
yarl==1.22.0
zipp==3.23.0
zope-interface==8.1.1
